requirements.txt
openai
python-dotenv
ftfy
pypdf
gradio
vllm
pillow